Make formatting functions methods of Translation objects
