+++ cms/locale/locale.py
@@ -33,6 +33,10 @@ from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 from __future__ import unicode_literals
+
+import copy
+
+import math
 from future.builtins.disabled import *
 from future.builtins import *
 import six
@@ -50,6 +54,7 @@ import babel.support
 import babel.units
 
 from cms import config
+from cmscommon.datetime import utc
 
 
 logger = logging.getLogger(__name__)
@@ -128,29 +133,160 @@ class Translation(object):
         else:
             return self.translation.ungettext(msgid1, msgid2, n)
 
-    def format_datetime(self, *args, **kwargs):
-        kwargs["locale"] = self.locale
-        return babel.dates.format_datetime(*args, **kwargs)
+    def format_datetime(self, dt, timezone):
+        """Return the date and time of dt.
 
-    def format_time(self, *args, **kwargs):
-        kwargs["locale"] = self.locale
-        return babel.dates.format_time(*args, **kwargs)
+        dt (datetime): a datetime.
+        timezone (tzinfo): the timezone the output should be in.
 
-    def format_timedelta(self, *args, **kwargs):
-        kwargs["locale"] = self.locale
-        return babel.dates.format_timedelta(*args, **kwargs)
+        return (str): the formatted date and time of the datetime.
 
-    def format_list(self, *args, **kwargs):
-        kwargs["locale"] = self.locale
-        return babel.lists.format_list(*args, **kwargs)
+        """
+        return babel.dates.format_datetime(dt, tzinfo=timezone,
+                                           locale=self.locale)
 
-    def format_decimal(self, *args, **kwargs):
-        kwargs["locale"] = self.locale
-        return babel.numbers.format_decimal(*args, **kwargs)
+    def format_time(self, dt, timezone):
+        """Return the time of dt.
 
-    def format_unit(self, *args, **kwargs):
-        kwargs["locale"] = self.locale
-        return babel.units.format_unit(*args, **kwargs)
+        dt (datetime): a datetime.
+        timezone (tzinfo): the timezone the output should be in.
+
+        return (str): the formatted time of the datetime.
+
+        """
+        return babel.dates.format_time(dt, tzinfo=timezone, locale=self.locale)
+
+    def format_datetime_smart(self, dt, now, timezone):
+        """Return dt formatted as '[date] time'.
+
+        Date is present in the output if it is not today.
+
+        dt (datetime): a datetime.
+        now (datetime): a datetime representing the moment in time at
+            which the previous parameter (dt) is being formatted.
+        timezone (tzinfo): the timezone the output should be in.
+
+        return (str): the formatted [date and] time of the datetime.
+
+        """
+        dt_date = dt.replace(tzinfo=utc).astimezone(timezone).date()
+        now_date = now.replace(tzinfo=utc).astimezone(timezone).date()
+        if dt_date == now_date:
+            return self.format_time(dt, timezone)
+        else:
+            return self.format_datetime(dt, timezone)
+
+    SECONDS_PER_HOUR = 3600
+    SECONDS_PER_MINUTE = 60
+
+    def format_timedelta(self, td):
+        """Return the timedelta formatted to high precision.
+
+        The result will be formatted as 'A days, B hours, C minutes and D
+        seconds', with components that would have a zero value removed. The
+        number of seconds has fractional digits, without trailing zeros.
+        Unlike Babel's built-in format_timedelta, no approximation or
+        rounding is performed.
+
+        td (timedelta): a timedelta.
+
+        return (string): the formatted timedelta.
+
+        """
+        td = abs(td)
+
+        res = []
+
+        if td.days > 0:
+            res.append(babel.units.format_unit(td.days, "duration-day",
+                                               locale=self.locale))
+
+        secs = td.seconds
+        if secs >= self.SECONDS_PER_HOUR:
+            res.append(babel.units.format_unit(secs // self.SECONDS_PER_HOUR,
+                                               "duration-hour",
+                                               locale=self.locale))
+            secs %= self.SECONDS_PER_HOUR
+        if secs >= self.SECONDS_PER_MINUTE:
+            res.append(babel.units.format_unit(secs // self.SECONDS_PER_MINUTE,
+                                               "duration-minute",
+                                               locale=self.locale))
+            secs %= self.SECONDS_PER_MINUTE
+        if secs > 0:
+            res.append(babel.units.format_unit(secs, "duration-second",
+                                               locale=self.locale))
+
+        if len(res) == 0:
+            res.append(babel.units.format_unit(0, "duration-second",
+                                               locale=self.locale))
+
+        return babel.lists.format_list(res, locale=self.locale)
+
+    def format_duration(self, d, length="short"):
+        """Format a duration in seconds.
+
+        Format the duration, usually of an operation performed in the
+        sandbox (compilation, evaluation, ...), given as a number of
+        seconds, always using a millisecond precision.
+
+        d (float): a duration, as a number of seconds.
+
+        returns (str): the formatted duration.
+
+        """
+        d = abs(d)
+
+        f = copy.copy(self.locale.decimal_formats[None])
+        f.frac_prec = (3, 3)
+        return babel.units.format_unit(d, "duration-second",
+                                       length=length, format=f, locale=self.locale)
+
+    PREFIX_FACTOR = 1000
+    SIZE_UNITS = ["byte", "kilobyte", "megabyte", "gigabyte", "terabyte"]
+
+    def format_size(self, n):
+        """Format the given number of bytes.
+
+        Format the size of a file, a memory allocation, etc. which is given
+        as a number of bytes. Use the most appropriate unit, from bytes up
+        to terabytes. Always use three significant digits, except when this
+        would mean:
+        - rounding the integral part (happens only for > 1000 terabytes),
+          in which case use more than three;
+        - showing sub-byte values (happens only for < 100 bytes), in which
+          case use less than three.
+
+        n (int): a size, as number of bytes.
+
+        return (str): the formatted size.
+
+        """
+        n = abs(n)
+
+        if n < self.PREFIX_FACTOR:
+            return babel.units.format_unit(round(n),
+                                           "digital-%s" % self.SIZE_UNITS[0],
+                                           length="short", locale=self.locale)
+        for unit in self.SIZE_UNITS[1:]:
+            n /= self.PREFIX_FACTOR
+            if n < self.PREFIX_FACTOR:
+                f = copy.copy(self.locale.decimal_formats[None])
+                d = max(int(math.ceil(math.log10(self.PREFIX_FACTOR / n))) - 1, 0)
+                f.frac_prec = (d, d)
+                return babel.units.format_unit(n, "digital-%s" % unit,
+                                               length="short", format=f, locale=self.locale)
+        return babel.units.format_unit(round(n), "digital-%s" % self.SIZE_UNITS[-1],
+                                       length="short", locale=self.locale)
+
+    def format_decimal(self, n):
+        """Format a (possibly decimal) number
+
+        n (float): the number to format.
+
+        returns (str): the formatted number.
+
+        """
+        return babel.numbers.format_decimal(n, locale=self.locale)
 
 
 DEFAULT_TRANSLATION = Translation("en")
