+++ cms/server/contest/formatting.py
@@ -49,172 +49,6 @@ N_("loading...")
 N_("unknown")
 
 
-def format_datetime(dt, timezone, translation=DEFAULT_TRANSLATION):
-    """Return the date and time of dt formatted as per locale.
-
-    dt (datetime): a datetime object.
-    timezone (tzinfo): the timezone the output should be in.
-    translation (Translation): the translation to use.
-
-    return (str): the date and time of dt, formatted using the given
-        locale.
-
-    """
-    return translation.format_datetime(dt, tzinfo=timezone)
-
-
-def format_time(dt, timezone, translation=DEFAULT_TRANSLATION):
-    """Return the time of dt formatted according to the given locale.
-
-    dt (datetime): a datetime object.
-    timezone (tzinfo): the timezone the output should be in.
-    translation (Translation): the translation to use.
-
-    return (str): the time of dt, formatted using the given locale.
-
-    """
-    return translation.format_time(dt, tzinfo=timezone)
-
-
-def format_datetime_smart(dt, now, timezone, translation=DEFAULT_TRANSLATION):
-    """Return dt formatted as '[date] time'.
-
-    Date is present in the output if it is not today.
-
-    dt (datetime): a datetime object to format.
-    now (datetime): a datetime object representing the moment in time
-        at which the previous parameter (dt) is being formatted.
-    timezone (tzinfo): the timezone the output should be in.
-    translation (Translation): the translation to use.
-
-    return (str): the [date and] time of dt, formatted using the given
-        locale.
-
-    """
-    dt_date = dt.replace(tzinfo=utc).astimezone(timezone).date()
-    now_date = now.replace(tzinfo=utc).astimezone(timezone).date()
-    if dt_date == now_date:
-        return format_time(dt, timezone, translation)
-    else:
-        return format_datetime(dt, timezone, translation)
-
-
-SECONDS_PER_HOUR = 3600
-SECONDS_PER_MINUTE = 60
-
-
-def format_timedelta(td, translation=DEFAULT_TRANSLATION):
-    """Return the timedelta formatted to high precision.
-
-    The result will be formatted as 'A days, B hours, C minutes and D
-    seconds', with components that would have a zero value removed. The
-    number of seconds has fractional digits, without trailing zeros.
-    Unlike Babel's built-in format_timedelta, no approximation or
-    rounding is performed.
-
-    td (timedelta): a timedelta.
-    translation (Translation): the translation to use.
-
-    return (string): the formatted timedelta.
-
-    """
-    td = abs(td)
-
-    res = []
-
-    if td.days > 0:
-        res.append(translation.format_unit(td.days, "duration-day"))
-
-    secs = td.seconds
-    if secs >= SECONDS_PER_HOUR:
-        res.append(translation.format_unit(secs // SECONDS_PER_HOUR,
-                                           "duration-hour"))
-        secs %= SECONDS_PER_HOUR
-    if secs >= SECONDS_PER_MINUTE:
-        res.append(translation.format_unit(secs // SECONDS_PER_MINUTE,
-                                           "duration-minute"))
-        secs %= SECONDS_PER_MINUTE
-    if secs > 0:
-        res.append(translation.format_unit(secs, "duration-second"))
-
-    if len(res) == 0:
-        res.append(translation.format_unit(0, "duration-second"))
-
-    return translation.format_list(res)
-
-
-def format_duration(d, length="short", translation=DEFAULT_TRANSLATION):
-    """Format a duration in seconds.
-
-    Format the duration, usually of an operation performed in the
-    sandbox (compilation, evaluation, ...), given as a number of
-    seconds, always using a millisecond precision.
-
-    d (float): a duration, as a number of seconds.
-    translation (Translation): the translation to use.
-
-    returns (str): the formatted duration.
-
-    """
-    d = abs(d)
-
-    f = copy.copy(translation.locale.decimal_formats[None])
-    f.frac_prec = (3, 3)
-    return translation.format_unit(d, "duration-second",
-                                   length=length, format=f)
-
-
-PREFIX_FACTOR = 1000
-SIZE_UNITS = ["byte", "kilobyte", "megabyte", "gigabyte", "terabyte"]
-
-
-def format_size(n, translation=DEFAULT_TRANSLATION):
-    """Format the given number of bytes.
-
-    Format the size of a file, a memory allocation, etc. which is given
-    as a number of bytes. Use the most appropriate unit, from bytes up
-    to terabytes. Always use three significant digits, except when this
-    would mean:
-    - rounding the integral part (happens only for > 1000 terabytes),
-      in which case use more than three;
-    - showing sub-byte values (happens only for < 100 bytes), in which
-      case use less than three.
-
-    n (int): a size, as number of bytes.
-    translation (Translation): the translation to use.
-
-    return (str): the formatted size.
-
-    """
-    n = abs(n)
-
-    if n < PREFIX_FACTOR:
-        return translation.format_unit(round(n), "digital-%s" % SIZE_UNITS[0],
-                                       length="short")
-    for unit in SIZE_UNITS[1:]:
-        n /= PREFIX_FACTOR
-        if n < PREFIX_FACTOR:
-            f = copy.copy(translation.locale.decimal_formats[None])
-            d = max(int(math.ceil(math.log10(PREFIX_FACTOR / n))) - 1, 0)
-            f.frac_prec = (d, d)
-            return translation.format_unit(n, "digital-%s" % unit,
-                                           length="short", format=f)
-    return translation.format_unit(round(n), "digital-%s" % SIZE_UNITS[-1],
-                                   length="short")
-
-
-def format_decimal(n, translation=DEFAULT_TRANSLATION):
-    """Format a (possibly decimal) number
-
-    n (float): the number to format.
-    translation (Translation): the translation to use.
-
-    returns (str): the formatted number.
-
-    """
-    return translation.format_decimal(n)
-
-
 def format_token_rules(tokens, t_type=None, translation=DEFAULT_TRANSLATION):
     """Return a human-readable string describing the given token rules
