+++ cms/server/contest/templates/communication.html
@@ -18,7 +18,7 @@
         {% else %}
         <h4 class="subject empty">{{ _("(no subject)") }}</h4>
         {% end %}
-        <span class="timestamp">{{ format_datetime_smart(msg.timestamp, now, timezone, translation=translation) }}</span>
+        <span class="timestamp">{{ translation.format_datetime_smart(msg.timestamp, now, timezone) }}</span>
         {% if len(msg.text) > 0 %}
         <div class="body">{{ msg.text }}</div>
         {% end %}
@@ -73,7 +73,7 @@
         {% else %}
         <h4 class="subject empty">{{ _("(no subject)") }}</h4>
         {% end %}
-        <span class="timestamp">{{ format_datetime_smart(msg.question_timestamp, now, timezone, translation=translation) }}</span>
+        <span class="timestamp">{{ translation.format_datetime_smart(msg.question_timestamp, now, timezone) }}</span>
         {% if len(msg.text) > 0 %}
         <div class="body">{{ msg.text }}</div>
         {% end %}
@@ -85,7 +85,7 @@
             {% else %}
         <h4 class="subject empty">{{ _("(no subject)") }}</h4>
             {% end %}
-        <span class="timestamp">{{ format_datetime_smart(msg.reply_timestamp, now, timezone, translation=translation) }}</span>
+        <span class="timestamp">{{ translation.format_datetime_smart(msg.reply_timestamp, now, timezone) }}</span>
             {% if len(msg.reply_text) > 0 %}
         <div class="body">{{ msg.reply_text }}</div>
             {% end %}
@@ -112,7 +112,7 @@
         {% else %}
         <h4 class="subject empty">{{ _("(no subject)") }}</h4>
         {% end %}
-        <span class="timestamp">{{ format_datetime_smart(msg.timestamp, now, timezone, translation=translation) }}</span>
+        <span class="timestamp">{{ translation.format_datetime_smart(msg.timestamp, now, timezone) }}</span>
         {% if len(msg.text) > 0 %}
         <div class="body">{{ msg.text }}</div>
         {% end %}
