+++ cms/server/contest/templates/contest.html
@@ -9,7 +9,7 @@
 {% import re %}
 {% import time %}
 {% import json %}
-{% from cms.server.contest.formatting import format_timedelta, format_time, format_datetime, format_datetime_smart, get_score_class, encode_for_url %}
+{% from cms.server.contest.formatting import get_score_class, encode_for_url %}
 {% from cms.grading import COMPILATION_MESSAGES, EVALUATION_MESSAGES %}
 {% from cms.grading.languagemanager import get_language %}
 {% from cms.db import SubmissionResult %}
