+++ cms/server/contest/templates/submission_details.html
@@ -1,7 +1,6 @@
 {% from six import iterkeys, itervalues, iteritems %}
 
 {% from cms.grading import format_status_text %}
-{% from cms.server.contest.formatting import format_duration, format_size %}
 
 {% if details is not None %}
 <div class="score_details">
@@ -23,7 +22,7 @@
 {% if sr.compilation_time is None %}
     {{ _("N/A") }}
 {% else %}
-    {{ format_duration(sr.compilation_time, translation=translation) }}
+    {{ translation.format_duration(sr.compilation_time) }}
 {% end %}
             </td>
         </tr>
@@ -32,7 +31,7 @@
 {% if sr.compilation_memory is None %}
     {{ _("N/A") }}
 {% else %}
-    {{ format_size(sr.compilation_memory, translation=translation) }}
+    {{ translation.format_size(sr.compilation_memory) }}
 {% end %}
             </td>
         </tr>
