+++ cms/server/contest/templates/task_description.html
@@ -1,7 +1,7 @@
 {% extends contest.html %}
 {% block core %}
 
-{% from cms.server.contest.formatting import format_token_rules, format_duration, format_size %}
+{% from cms.server.contest.formatting import format_token_rules %}
 {% from cms.grading.tasktypes import get_task_type %}
 
 {% if actual_phase == 0 or actual_phase == 3 or current_user.unrestricted %}
@@ -99,13 +99,13 @@
 {% if task.active_dataset.time_limit is not None %}
         <tr>
             <th>{{ _("Time limit") }}</th>
-            <td colspan="2">{{ format_duration(task.active_dataset.time_limit, length="long", translation=translation) }}</td>
+            <td colspan="2">{{ translation.format_duration(task.active_dataset.time_limit, length="long") }}</td>
         </tr>
 {% end %}
 {% if task.active_dataset.memory_limit is not None %}
         <tr>
             <th>{{ _("Memory limit") }}</th>
-            <td colspan="2">{{ format_size(task.active_dataset.memory_limit * 1024 * 1024, translation=translation) }}</td>
+            <td colspan="2">{{ translation.format_size(task.active_dataset.memory_limit * 1024 * 1024) }}</td>
         </tr>
 {% end %}
 {% set compilation_commands = task_type.get_compilation_commands([x.filename for x in task.submission_format]) %}
@@ -188,7 +188,7 @@
             {% end %}
                     <span class="first_line">
                         <span class="name">{{ filename }}</span>
-                        <span class="size">{{ format_size(file_size, translation=translation) }}</span>
+                        <span class="size">{{ translation.format_size(file_size) }}</span>
                     </span>
             {% if type_icon is not None %}
                     <span class="type">{{ _(type_name) }}</span>
